refactor(encoding): split optimal-parse DP into hc/optimal.rs
Move the binary-tree optimal parser out of hc/generator.rs into hc/optimal.rs:
the build_optimal_plan DP + its per-CPU-tier kernels, the candidate-collection /
price-set body macros, and the btlazy2 cost helpers, as a second
`impl HcMatchGenerator` block over the same matcher. Both files are hc/
submodules so no path rewrite was needed; output is byte-identical.

The DP body macros reference opt types unqualified, which rustc's unused-import
lint cannot see (so it false-flags every such import) -- handled by a documented
module-level suppression in optimal.rs; generator.rs's now-redundant (qualified)
imports were removed by hand. hc/generator.rs: 3918 -> 1691.

Part of #441.
